+++ pkg/client/controllerPing.go
@@ -10,12 +10,12 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 )
 
-func (c *client) getPing(ctx context.Context, host string) (*types.Ping, error) {
-	var ping *types.Ping
+func (c *client) getControllerPing(ctx context.Context, host string) (*types.ControllerPing, error) {
+	var ping *types.ControllerPing
 	var err error
 	log := log.FromContext(ctx)
 	for i := c.clientUse; i < len(c.clients); i++ {
-		ping, err = c.clients[i].GetPing(ctx, host)
+		ping, err = c.clients[i].GetControllerPing(ctx, host)
 		if err == nil {
 			c.clientUse = i
 			break
@@ -26,16 +26,16 @@ func (c *client) getPing(ctx context.Context, host string) (*types.Ping, error) 
 	if err != nil {
 		return nil, err
 	}
-	log.V(1).Info("GetPing()", "ping", ping)
+	log.V(1).Info("GetControllerPing()", "ping", ping)
 	return ping, nil
 }
 
-func (c *client) listPing(ctx context.Context) (*types.PingList, error) {
-	var pingList *types.PingList
+func (c *client) listControllerPing(ctx context.Context) (*types.ControllerPingList, error) {
+	var pingList *types.ControllerPingList
 	var err error
 	log := log.FromContext(ctx)
 	for i := c.clientUse; i < len(c.clients); i++ {
-		pingList, err = c.clients[i].ListPing(ctx)
+		pingList, err = c.clients[i].ListControllerPing(ctx)
 		if err == nil {
 			c.clientUse = i
 			break
@@ -46,6 +46,6 @@ func (c *client) listPing(ctx context.Context) (*types.PingList, error) {
 	if err != nil {
 		return nil, err
 	}
-	log.V(1).Info("ListPing()", "size", len(pingList.Items), "pingList", pingList)
+	log.V(1).Info("ListControllerPing()", "size", len(pingList.Items), "controllerPingList", pingList)
 	return pingList, nil
 }